pyproject.toml
[project]
name = "ai-cv-parsing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "nomad-at-code", email = "gitea.pusher020@passinbox.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "pymupdf>=1.26.4",
    "python-magic>=0.4.27",
    "textract>=1.6.5",
]

[project.scripts]
ai-cv-parsing = "ai_cv_parsing:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
